# Avoid AliasAnalysis for Hello_World.enso