--- a/engine/runtime-compiler/src/main/scala/org/enso/compiler/phase/ImportResolver.scala
+++ b/engine/runtime-compiler/src/main/scala/org/enso/compiler/phase/ImportResolver.scala
@@ -89,9 +89,9 @@ final class ImportResolver(compiler: Compiler) extends ImportResolverForIR {
         val newImportIRs =
           importedModules.map(_._1) ++ syntheticImports.map(_._1)
 
-        currentLocal.resolvedImports_(
+        currentLocal.resolvedImports =
           resolvedImports ++ resolvedSyntheticImports
-        )
+
         val newIr = ir.copy(imports = newImportIRs)
         context.updateModule(
           current,