--- a/engine/runtime-compiler/src/main/scala/org/enso/compiler/phase/exports/ExportsResolution.scala
+++ b/engine/runtime-compiler/src/main/scala/org/enso/compiler/phase/exports/ExportsResolution.scala
@@ -179,7 +179,7 @@ class ExportsResolution(private val context: CompilerContext) {
           }
           (symbolName, resolvedNames)
         }
-      bindings.exportedSymbols_(newSymbols)
+      bindings.exportedSymbols = newSymbols
     }
   }