### Pull Request Description

Fixes #10843 in an even better way then #10906. At the end of work on #10906 it has been found that `AliasAnalysis` is still used for `Hello_World.enso`. Mostly because when loading its `.ir` cache, we cannot resolve its `Standard.Base.IO` import as nobody loaded the `Standard.Base` package into the system yet. This PR fixes the problem by __delaying resolution of `BindingsMap` internals until one of its getters is called__.

At the end `Hello_World` loads its `.ir` caches properly without invoking any `AliasAnalysis` functionality as verified by checking the logs.

### Checklist

Please ensure that the following checklist has been satisfied before submitting the PR:

- [x] All code follows the
      [Scala](https://github.com/enso-org/enso/blob/develop/docs/style-guide/scala.md),
      [Java](https://github.com/enso-org/enso/blob/develop/docs/style-guide/java.md),
- [x] Unit tests have been written where possible.
- [x] Benchmarks are improved - [1st run](https://github.com/enso-org/enso/actions/runs/10733289249) - results [are good](https://github.com/enso-org/enso/pull/10996#issuecomment-2333708114)

### Steps to Reproduce

The ultimate way to verify the behavior is in debugger. Place a breakpoint to `Graph.scala:12` - it should get trigger only once when `LocalScope.empty` is being created. Possibly also for the second time when execution is over and #10842 kicks in (this second case can be disabled by using `--repl` for example).

To reproduce from CLI run twice (first run generates the `.ir` cache) and then checks the logs:
```bash
enso$ ./built-distribution/enso-engine-*/enso-*/bin/enso --run test/Benchmarks/src/Startup/Hello_World.enso
enso$ ./built-distribution/enso-engine-*/enso-*/bin/enso --run test/Benchmarks/src/Startup/Hello_World.enso --log-level debug | grep Hello
```
prior to this PR the log contains:
```
[enso.org.enso.interpreter.runtime.SerializationPool] Unable to load a cache for module [Hello_World].
[enso.org.enso.interpreter.runtime.SerializationPool] Deserializing module Hello_World from IR file: false
[enso.org.enso.compiler.Compiler] Loading module [Hello_World] from source.
[enso.org.enso.compiler.Compiler] Generating code for module [Hello_World].
```
with this PR the log contains:
```
[enso.org.enso.compiler.Compiler] Parsing module [Hello_World].
[enso.org.enso.interpreter.runtime.SerializationPool] Restored IR from cache for module [Hello_World] at stage [AFTER_STATIC_PASSES].
[enso.org.enso.interpreter.runtime.SerializationPool] Deserializing module Hello_World from IR file: true
[enso.org.enso.compiler.Compiler] Generating code for module [Hello_World].
```