@JaroslavTulach 2024-09-06
There is **5% speedup** on `Import_World` startup benchmark:

![Import_World](https://github.com/user-attachments/assets/6f46b294-8e1c-4fc3-9468-e8299f1ec5d5)

obviously `Hello_World` startup shows lower improvements (around 100ms) and `empty_startup` shows none (as there is no import).