@JaroslavTulach 2024-09-06
I don't trust local benchmarking, but per @Akirathan request I did one:

### Old Run
```
Benchmark                     Mode  Cnt     Score      Error  
Startup.empty_startup         avgt    3  2438.773 ± 2213.236  
Startup.hello_world_startup   avgt    3  3550.110 ±   65.246  
Startup.import_world_startup  avgt    3  4118.481 ±  938.980 
```
### New run
```
Benchmark                     Mode  Cnt     Score      Error 
Startup.empty_startup         avgt    3  2468.019 ± 1609.600
Startup.hello_world_startup   avgt    3  3323.337 ±  811.669
Startup.import_world_startup  avgt    3  4097.680 ±  569.984
```
The benchmark fluctuates a lot, but there seems to be no slowdown and possibly there is some speedup.